add _create_unverified_context in download_ulr
