+++ torch_points3d/datasets/classification/modelnet.py
@@ -1,9 +1,7 @@
 import os.path as osp
-from six.moves import urllib
 import os
 import shutil
 import torch
-import ssl
 
 from torch_geometric.datasets import ModelNet
 from torch_geometric.data import DataLoader, InMemoryDataset, extract_zip, Data
@@ -12,37 +10,7 @@ from torch_geometric.io import read_txt_array
 
 from torch_points3d.datasets.base_dataset import BaseDataset
 from torch_points3d.metrics.classification_tracker import ClassificationTracker
-
-def download_url(url, folder, log=True):
-    r"""Downloads the content of an URL to a specific folder.
-
-    Args:
-        url (string): The url.
-        folder (string): The folder.
-        log (bool, optional): If :obj:`False`, will not print anything to the
-            console. (default: :obj:`True`)
-    """
-
-    filename = url.rpartition('/')[2]
-    path = osp.join(folder, filename)
-
-    if osp.exists(path):  # pragma: no cover
-        if log:
-            print('Using exist file', filename)
-        return path
-
-    if log:
-        print('Downloading', url)
-
-    os.makedirs(folder)
-    context = ssl._create_unverified_context()
-    data = urllib.request.urlopen(url, context=context)
-
-    with open(path, 'wb') as f:
-        f.write(data.read())
-
-    return path
-
+from torch_points3d.utils.download import download_url
 
 class SampledModelNet(InMemoryDataset):
     r"""The ModelNet10/40 dataset from the `"3D ShapeNets: A Deep
@@ -100,7 +68,6 @@ class SampledModelNet(InMemoryDataset):
         return ["training_{}.pt".format(self.name), "test_{}.pt".format(self.name)]
 
     def download(self):
-        return
         path = download_url(self.url, self.root)
         extract_zip(path, self.root)
         os.unlink(path)
